requirements.txt
agno
python-telegram-bot
python-dotenv
openai
requests
sqlalchemy
aiosqlite
psycopg[binary]
psycopg2-binary
mcp
apscheduler
croniter
fpdf
edge-tts
tavily-python
